refactor: split data.py into domain query modules (codebase-health Phase 04)
Split the 2,865-line shitty_ui/data.py monolith into 5 focused modules
under shitty_ui/data/ package for improved developer experience:

- data/base.py: shared infrastructure (execute_query, ttl_cache, logger)
- data/signal_queries.py: 13 signal loading, feed, and filtering functions
- data/performance_queries.py: 21 KPI, accuracy, P&L, and backtesting functions
- data/asset_queries.py: 9 screener, sparkline, and asset stats functions
- data/insight_queries.py: 1 dynamic insight card generation function
- data/__init__.py: re-exports all 44 public functions for zero breaking changes

All submodules use `import data.base as _base` pattern to ensure
@patch("data.base.execute_query") works correctly in tests. Updated
155 patch targets across 3 test files. Added 7 re-export contract tests.

No logic, query, or API changes. Pure structural refactoring.
